+++ CHANGES.md
@@ -20,6 +20,8 @@ Release Notes.
 * Fixed the invalid issue in the isInterface method in PluginFinder.
 * Fix the opentracing toolkit SPI config
 * Improve 4x performance of ContextManagerExtendService.createTraceContext()
+* Add a plugin that supports the Solon framework.
+
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/213?closed=1)